refactor: extract match starter class from method
