+++ aiarena/core/services/matches.py
@@ -28,6 +28,7 @@ from aiarena.core.models import (
 )
 from aiarena.core.services import Bots
 from aiarena.core.services.competitions import Competitions
+from aiarena.core.services.internal.match_starter import MatchStarter
 from aiarena.core.services.internal.matches import CancelResult, cancel, create
 
 
@@ -64,59 +65,6 @@ class Matches:
                 match.round.update_if_completed()
 
     @staticmethod
-    def __start_match(match: Match, arenaclient: ArenaClient) -> bool:
-        if Matches.__requires_trusted_arenaclient(match) and not arenaclient.trusted:
-            return False
-
-        match.lock_me()  # lock self to avoid race conditions
-
-        if match.is_already_started:
-            logger.warning(f"Match {match.id} failed to start unexpectedly as it was already started.")
-            return False
-
-        # Avoid starting a match when a participant is not available
-        participations = MatchParticipation.objects.raw(
-            """
-            SELECT cm.id FROM core_matchparticipation cm
-            where ((cm.use_bot_data or cm.update_bot_data) or cm.bot_id not in (
-                    select bot_id
-                    from core_matchparticipation
-                    inner join core_match m on core_matchparticipation.match_id = m.id
-                    left join core_result cr on m.result_id = cr.id
-                    where m.started is not null
-                    and cr.type is null
-                    and (core_matchparticipation.use_bot_data  or core_matchparticipation.update_bot_data)       
-                    and m.id != %s 
-                )) and match_id = %s
-                """,
-            (match.id, match.id),
-        )
-
-        if len(participations) < 2:
-            # Todo: Commented out to avoid log spam. This used to be a last second sanity check.
-            # Todo: Investigate whether it is still the case or whether this is no longer considered a system fault
-            # Todo: worthy of a warning message being logged.
-            # logger.warning(f"Match {match.id} failed to start unexpectedly"
-            #                f" because one of the participants was not available.")
-            return False
-
-        # If all checks pass, start the match
-        match.started = match.first_started = timezone.now()
-        match.assigned_to = arenaclient
-        match.save()
-        return True
-
-    @staticmethod
-    def __requires_trusted_arenaclient(match):
-        require_trusted_arenaclient = match.require_trusted_arenaclient or (
-            not match.participant1.bot.bot_zip_publicly_downloadable
-            or not match.participant2.bot.bot_zip_publicly_downloadable
-            or not match.participant1.bot.bot_data_publicly_downloadable
-            or not match.participant2.bot.bot_data_publicly_downloadable
-        )
-        return require_trusted_arenaclient
-
-    @staticmethod
     def attempt_to_start_a_requested_match(requesting_ac: ArenaClient):
         # Try get a requested match
         # Do we want trusted clients to run games not requiring trusted clients?
@@ -135,7 +83,7 @@ class Matches:
     @staticmethod
     def _start_and_return_a_match(requesting_ac: ArenaClient, matches):
         for match in matches:
-            if Matches.__start_match(match, requesting_ac):
+            if MatchStarter.start(match, requesting_ac):
                 return match
         return None  # No match was able to start
